Add new script omitdynamics
This is useful to remove the visibility of dynamics in a lilypond
input file. Thus a score can be entered with all the dynamics in a part
so that the midi output has all the right dynamics, but the score itself
can use Dynamics contexts to print only the parts which are required.